+++ airless/hook/notification/slack.py
@@ -18,7 +18,7 @@ class SlackHook(BaseHook):
             'Authorization': f'Bearer {self.token}'
         }
 
-    def send(self, channel, message=None, blocks=None, thread_ts=None, reply_broadcast=False):
+    def send(self, channel, message=None, blocks=None, thread_ts=None, reply_broadcast=False, attachments=None):
 
         data = {
             'channel': channel,
@@ -32,6 +32,9 @@ class SlackHook(BaseHook):
         if blocks:
             data['blocks'] = blocks
 
+        if attachments:
+            data['attachments'] = attachments
+
         if thread_ts:
             data['thread_ts'] = thread_ts
             data['reply_broadcast'] = reply_broadcast
@@ -43,6 +46,7 @@ class SlackHook(BaseHook):
             timeout=10
         )
         response.raise_for_status()
+        return response.json()
 
     def react(self, channel, reaction, ts):
         data = {
@@ -57,3 +61,4 @@ class SlackHook(BaseHook):
             timeout=10
         )
         response.raise_for_status()
+        return response.json()